+++ src/bsp.h
@@ -53,7 +53,7 @@ void bsp_destroy(bsp_t *b);
 
 // A convenience method to create a new bsp of the given
 // size, copy across all points, and destroy the old bsp.
-bsp_t *bsp_change_size(bsp_t *b, int new_size);
+bsp_t *bsp_change_size(bsp_t *b, double new_size);
 
 // Adds a point to the tree
 void bsp_add_point(bsp_t *b, double x, double y);
